main: prioritize switches with cpp directives and/or description
There are few switches that are defined at different places. For instance, the
switch --enable-gpu-service-logging is defined in two files:
* gpu/command_buffer/service/gpu_switches.cc [1]
* ui/gl/gl_switches.cc [2]

An extra note is visible in gpu/command_buffer/service/gpu_switches.cc:

  > Note: This is the same switch as the one in gl_switches.cc. It's defined
  > here again to avoid dependencies between dlls [3]

At least, the reason why this option is defined twice is explicit --though it
needs a human to read and understand the comment

Another example is --nacl-loader, defined in two different files as well:
* sandbox/policy/switches.cc [4]
* components/nacl/common/nacl_switches [5]

Again, an extra comment explains it:

  // Flags spied upon from other layers. [6]

Here is another interesting example. --relauncher is defined in the following
files:
* chrome/common/chrome_switches.cc [7]
* sandbox/policy/switches.cc [8]

Though, --relauncher is inside a cpp directive "#if BUILDFLAG(IS_MAC) in [7] but
not in [8]. The cpp directive is not emphasized by peter.sh [9], though it
should since the cpp condition does exist.

To add more precision in the mining of switches, we decide to prioritize the
switches mined from a file where we have more precision such as comment
description or cpp directives.

[1] https://github.com/chromium/chromium/blob/main/gpu/command_buffer/service/gpu_switches.cc#L30
[2] https://github.com/chromium/chromium/blob/main/ui/gl/gl_switches.cc#L69
[3] https://github.com/chromium/chromium/blob/main/gpu/command_buffer/service/gpu_switches.cc#L27
[4] https://github.com/chromium/chromium/blob/main/sandbox/policy/switches.cc#L132
[5] https://github.com/chromium/chromium/blob/main/components/nacl/common/nacl_switches.cc#L36
[6] https://github.com/chromium/chromium/blob/main/sandbox/policy/switches.cc#L129C1-L129C39
[7] https://github.com/chromium/chromium/blob/main/chrome/common/chrome_switches.cc#L788
[8] https://github.com/chromium/chromium/blob/main/sandbox/policy/switches.cc#L137
[9] https://peter.sh/experiments/chromium-command-line-switches/?date=2023-09-22#relauncher